Add sensors

+++ components/daikin_rotex_uart/daikin_rotex_uart.h
@@ -23,7 +23,7 @@ class DaikinRotexUARTComponent: public Component, public uart::UARTDevice {
         uint8_t registryID;
         uint8_t offset;
         bool isSigned;
-        int dataSize;
+        uint8_t dataSize;
         Endian endian;
         double divider;
     };
